--- lib/features/authentication/controllers/otp_controller.dart
@@ -1,11 +1,0 @@
-// import 'package:get/get.dart';
-// import 'package:id_generator/pages/student_home.dart';
-
-// import 'package:id_generator/src/repository/authentication_repository/authentication_repository.dart';
-
-// class OTPController extends GetxController {
-//   void verifyOTP(String otp) async {
-//     var isVerified = await AuthenticationRepository.instance.verifyOtp(otp);
-//     isVerified ? Get.offAll(const StudentHome()) : Get.back();
-//   }
-// }
