+++ lib/pages/generate_qr_code.dart
