+++ lib/pages/qr_scanner.dart
