+++ lib/pages/student_qr.dart
@@ -6,7 +6,7 @@ import 'package:id_generator/pages/verify_otp.dart';
 import 'package:qr_flutter/qr_flutter.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
-import '../features/authentication/screens/login.dart';
+import 'login.dart';
 
 class StudentQR extends StatelessWidget {
   final String data;
